requirements.txt
django==3.1.2
gunicorn==20.0.4
psycopg2==2.8.5
djangorestframework==3.1.2
dj-database-url==0.5.0
whitenoise==5.2.0